# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Game Theory Research Lab API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /experiments/{experimentId}/analysis:
    get:
      operationId: getAnalysis
      tags: [analyses]
      summary: Get statistical analysis for an experiment
      parameters:
        - name: experimentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Analysis result
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "404":
          description: Analysis not found
    post:
      operationId: createAnalysis
      tags: [analyses]
      summary: Run statistical analysis for a completed experiment
      parameters:
        - name: experimentId
          in: path
          required: true
          schema:
            type: integer
      responses:
        "201":
          description: Analysis created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Analysis"
        "400":
          description: Experiment not completed
  /analyses/aggregate:
    get:
      operationId: getAggregateAnalysis
      tags: [analyses]
      summary: Aggregate v2 metrics across analyzed experiments (mean, sd, 95% CI)
      parameters:
        - name: gameId
          in: query
          schema:
            type: integer
        - name: player1StrategyId
          in: query
          schema:
            type: integer
        - name: player2StrategyId
          in: query
          schema:
            type: integer
        - name: batchLabel
          in: query
          schema:
            type: string
      responses:
        "200":
          description: Aggregated metrics
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/AggregateResponse"

  # ── Claims ───────────────────────────────────────────────────────────────
  /claims:
    get:
      operationId: listClaims
      tags: [claims]
      summary: List all research claims
      parameters:
        - name: gameId
          in: query
          schema:
            type: integer
        - name: status
          in: query
          schema:
            type: string
            enum: [hypothesis, supported, refuted, inconclusive, untested]
      responses:
        "200":
          description: List of claims
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Claim"
    post:
      operationId: createClaim
      tags: [claims]
      summary: Create a research claim
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimInput"
      responses:
        "201":
          description: Created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "400":
          description: Invalid input
  /claims/{id}:
    get:
      operationId: getClaim
      tags: [claims]
      summary: Get a claim by ID
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Claim found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          description: Not found
    patch:
      operationId: updateClaim
      tags: [claims]
      summary: Update a claim's status or evidence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/ClaimUpdate"
      responses:
        "200":
          description: Updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "404":
          description: Not found
    delete:
      operationId: deleteClaim
      tags: [claims]
      summary: Delete a claim
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
        "404":
          description: Not found

  # ── Papers ───────────────────────────────────────────────────────────────
...
  /claims/{id}/adjudicate:
    post:
      operationId: adjudicateClaim
      tags: [claims]
      summary: Mechanically adjudicate a single claim against matching experimental evidence
      parameters:
        - name: id
          in: path
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Updated claim with adjudication record
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Claim"
        "400":
          description: Claim has no structured predicate
        "404":
          description: Not found
...
components:
  schemas:
...
    Analysis:
      type: object
      required: [id, experimentId, nashEquilibriumRate, player1CooperationRate, player2CooperationRate, player1AvgPayoff, player2AvgPayoff, theoreticalPlayer1Payoff, theoreticalPlayer2Payoff, player1PayoffDeviation, player2PayoffDeviation, mutualCooperationRate, mutualDefectionRate, mixedOutcomeRate, summary, createdAt]
      properties:
        id:
          type: integer
        experimentId:
          type: integer
        nashEquilibriumRate:
          type: number
          description: Fraction of rounds landing on Nash equilibrium
        player1CooperationRate:
          type: number
        player2CooperationRate:
          type: number
        player1AvgPayoff:
          type: number
        player2AvgPayoff:
          type: number
        theoreticalPlayer1Payoff:
          type: number
          description: Expected payoff if players played Nash every round
        theoreticalPlayer2Payoff:
          type: number
        player1PayoffDeviation:
          type: number
          description: Percentage deviation from theoretical Nash payoff
        player2PayoffDeviation:
          type: number
        mutualCooperationRate:
          type: number
        mutualDefectionRate:
          type: number
        mixedOutcomeRate:
          type: number
        roundByRoundJson:
          type: ["string", "null"]
          description: JSON array of cumulative stats per round
        analysisVersion:
          type: integer
          description: 1 = legacy metrics; 2 = per-game-class metrics in metricsJson
        metricsJson:
          type: ["string", "null"]
          description: JSON-encoded v2 metrics keyed by metric name (per game class)
        summary:
          type: string
          description: Human-readable analysis narrative
        createdAt:
          type: string
          format: date-time

    # ── Claim ─────────────────────────────────────────────────────────────
    ClaimInput:
      type: object
      required: [title, statement, gameId]
      properties:
        title:
          type: string
        statement:
          type: string
          description: The full claim as a testable proposition
        gameId:
          type: integer
        strategyId:
          type: ["integer", "null"]
        analysisId:
          type: ["integer", "null"]
        evidenceSummary:
          type: string
        predicateJson:
          type: string
          description: JSON-encoded structured predicate for mechanical adjudication
    ClaimUpdate:
      type: object
      description: >-
        Claim status is intentionally NOT updatable here — verdicts are
        machine-assigned via the adjudication endpoints only.
      properties:
        title:
          type: string
        statement:
          type: string
        evidenceSummary:
          type: string
        linkedAnalysisId:
          type: ["integer", "null"]
        predicateJson:
          type: string
          description: JSON-encoded structured predicate for mechanical adjudication
    Claim:
      type: object
      required: [id, title, statement, gameId, status, createdAt]
      properties:
        id:
          type: integer
        title:
          type: string
        statement:
          type: string
        gameId:
          type: integer
        gameName:
          type: ["string", "null"]
        strategyId:
          type: ["integer", "null"]
        strategyName:
          type: ["string", "null"]
        analysisId:
          type: ["integer", "null"]
        status:
          type: string
          enum: [hypothesis, supported, refuted, inconclusive, untested]
        evidenceSummary:
          type: ["string", "null"]
        predicateJson:
          type: ["string", "null"]
          description: JSON-encoded structured predicate for mechanical adjudication
        adjudicationJson:
          type: ["string", "null"]
          description: JSON-encoded latest adjudication record (evidence, stats, verdict)
        adjudicatedAt:
          type: ["string", "null"]
          format: date-time
        createdAt:
          type: string
          format: date-time
        updatedAt:
          type: string
          format: date-time

    # ── Batch & Aggregation ───────────────────────────────────────────────
...
    MetricAggregate:
      type: object
      required: [name, n, mean]
      properties:
        name:
          type: string
        n:
          type: integer
        mean:
          type: number
        sd:
          type: ["number", "null"]
          description: Sample standard deviation; null when n < 2
        ciLow:
          type: ["number", "null"]
          description: Lower bound of 95% t-interval; null when n < 2
        ciHigh:
          type: ["number", "null"]
    AggregateResponse:
      type: object
      required: [n, metrics]
      properties:
        n:
          type: integer
          description: Number of analyzed experiments matching the filter
        gameId:
          type: ["integer", "null"]
        player1StrategyId:
          type: ["integer", "null"]
        player2StrategyId:
          type: ["integer", "null"]
        batchLabel:
          type: ["string", "null"]
        metrics:
          type: array
          items:
            $ref: "#/components/schemas/MetricAggregate"
...