# lib/api-zod/src/generated/types/analysis.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Game Theory Research Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface Analysis {
  id: number;
  experimentId: number;
  /** Fraction of rounds landing on Nash equilibrium */
  nashEquilibriumRate: number;
  player1CooperationRate: number;
  player2CooperationRate: number;
  player1AvgPayoff: number;
  player2AvgPayoff: number;
  /** Expected payoff if players played Nash every round */
  theoreticalPlayer1Payoff: number;
  theoreticalPlayer2Payoff: number;
  /** Percentage deviation from theoretical Nash payoff */
  player1PayoffDeviation: number;
  player2PayoffDeviation: number;
  mutualCooperationRate: number;
  mutualDefectionRate: number;
  mixedOutcomeRate: number;
  /**
     * JSON array of cumulative stats per round
     * @nullable
     */
  roundByRoundJson?: string | null;
  /** 1 = legacy metrics; 2 = per-game-class metrics in metricsJson */
  analysisVersion?: number;
  /**
     * JSON-encoded v2 metrics keyed by metric name (per game class)
     * @nullable
     */
  metricsJson?: string | null;
  /** Human-readable analysis narrative */
  summary: string;
  createdAt: Date;
}