# lib/api-zod/src/generated/types/claimInput.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Game Theory Research Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface ClaimInput {
  title: string;
  /** The full claim as a testable proposition */
  statement: string;
  gameId: number;
  /** @nullable */
  strategyId?: number | null;
  /** @nullable */
  analysisId?: number | null;
  evidenceSummary?: string;
  /** JSON-encoded structured predicate for mechanical adjudication */
  predicateJson?: string;
}