# lib/api-zod/src/generated/types/metricAggregate.ts
/**
 * Generated by orval v8.22.0 🍺
 * Do not edit manually.
 * Api
 * Game Theory Research Lab API
 * OpenAPI spec version: 0.1.0
 */

export interface MetricAggregate {
  name: string;
  n: number;
  mean: number;
  /**
     * Sample standard deviation; null when n < 2
     * @nullable
     */
  sd?: number | null;
  /**
     * Lower bound of 95% t-interval; null when n < 2
     * @nullable
     */
  ciLow?: number | null;
  /** @nullable */
  ciHigh?: number | null;
}